Create individual files for tab components
