+++ src/App.scss
@@ -1,0 +1,3 @@
+//How do I use the color scheme?
+
+@import 'colors';
