+++ src/_colors.scss
@@ -1,0 +1,4 @@
+$purple: #d9d7f1;
+$yellow: #fffdde;
+$green: #e7fbbe;
+$red: #ffcbcb;
