+++ src/index.js
@@ -1,9 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
+import './index.scss';
 import App from './App';
 import 'bootstrap/dist/css/bootstrap.min.css';
-
 
 ReactDOM.render(
   <React.StrictMode>
