+++ src/tabs/ContactTab.jsx
@@ -1,0 +1,19 @@
+import React from 'react';
+
+const ContactTab = () => {
+  return(    
+    <>
+      <div>
+        Email:   echoalice0@gmail.com
+      </div><br></br>
+      <div>
+        Twitter:  <a href="https://twitter.com/echo__alice">https://twitter.com/echo__alice</a>
+      </div><br></br>
+      <div>
+        Github:  <a href="https://github.com/EchoAlice">https://github.com/EchoAlice</a>
+      </div>
+    </>
+  );
+}
+
+export default ContactTab;
