+++ src/tabs/HomeTab.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const HomeTab = () => {
+  return(    
+    <a href='https://giphy.com/clips/hamlet-jJjb9AUHOiP3nJJMdy'rel="noopener noreferrer" target="_blank">Hello!</a>
+  );
+}
+
+export default HomeTab;
