+++ src/tabs/ProfileTab.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const ProfileTab = () => {
+  return(    
+    <div>
+      Insert Pictures and link to podcast
+    </div>
+  );
+}
+
+export default ProfileTab;
