+++ src/tabs/ProjectsTab.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const ProjectsTab = () => {
+  return(    
+    <div>
+      Insert Merkle Tree display of script
+    </div>
+  );
+}
+
+export default ProjectsTab;
